requirements.txt
fastapi==0.92.0
python-multipart==0.0.5
uvicorn==0.20.0
motor~=3.1.1
pydantic~=1.10.5
pymongo~=4.3.3
kafka-python==2.0.2
exif~=1.6.0